Write test for UI
