+++ src/App.js
@@ -1,7 +1,7 @@
 import { Provider } from 'react-redux';
 import { Route, Switch } from 'react-router-dom';
-import DisplayData from './compoenets/display';
-import CovidDataRender from './compoenets/renderData';
+import DisplayData from './compoenets/continent';
+import CovidDataRender from './compoenets/renderContinent';
 import CountryDislay from './compoenets/country';
 import store from './redux/configureStore';
 import NavBar from './compoenets/navbar';
