+++ src/redux/configureStore.js
@@ -1,9 +1,9 @@
 import { createStore, combineReducers, applyMiddleware } from 'redux';
 import logger from 'redux-logger';
 import thunk from 'redux-thunk';
-import covidReducer from './metrics/metrics';
+import covidReducer from './continent/continent';
 import countryReducer from './country/country';
-import headingReducer from './continent/continent';
+import headingReducer from './continentName/continentName';
 
 const reducer = combineReducers({
   covidReducer,
